% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allSweptAreaspower.R
\docType{data}
\name{allSweptAreaspower}
\alias{allSweptAreaspower}
\title{Windmill data}
\format{
a data from with 200 rows and 4 columns representing different wind speeds
\itemize{
 \item data wind speed m/s
 \item data sweep area m^2
 \item power data kWh
 }
 @source Charles Louis Wilson IV
}
\usage{
allSweptAreaspower
}
\description{
Data generated to be used with function windmill_power
}
\keyword{datasets}
