% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windmill_power.R
\name{windmill_power}
\alias{windmill_power}
\title{Compute windmill power}
\usage{
windmill_power(
  p = 1.2,
  A,
  Cp = 0.35,
  V,
  Ng = 0.3,
  Nb = 0.5,
  price = 0.12,
  hidePrice,
  hidePower
)
}
\arguments{
\item{p}{is air density}

\item{A}{is rotor swept area m^2}

\item{Cp}{is power coefficient}

\item{V}{is wind velocity m/s}

\item{Ng}{is generator efficiency}

\item{Nb}{is gearbox efficiency}

\item{hidePrice}{if TRUE returns only power}

\item{hidePower}{if TRUE returns only price}
}
\value{
power output in kWh

revenue per second from windwill
}
\description{
This function computes the kWh power output of windmills and the revenue they generate per second
}
